backend/requirements.txt
# Enterprise Multi-Agent Financial Reporting Platform — backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
azure-identity>=1.19.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.55.0
numpy>=1.26.0
aiosqlite>=0.20.0
